+++ src/lctrack/service.py
@@ -1,21 +1,15 @@
-from pathlib import Path
 import random
 import sqlite3
+from collections.abc import Callable
 from datetime import datetime
-from typing import Callable
+from typing import Any
 
-import github
-import keyring
-from github.AuthenticatedUser import AuthenticatedUser
-from github.NamedUser import NamedUser
-from github.Repository import Repository
-
-import git
-from . import backup
-from .constants import BACKUP_EVENT_LOG, BACKUP_REPO_DIR, DIFF_COLOUR, LOCAL_EVENT_LOG, RESET, TMP_EVENT_LOG, YELLOW
+from lctrack import lc_client
 
 from . import access
-from .ds import AddEntryEvent, BaseEvent, Entry, Problem, RmEntryEvent
+from .constants import DIFF_COLOUR, RESET, YELLOW
+from .ds import DIFF_TO_INT, AddEntryEvent, Entry, Problem, RmEntryEvent
+
 
 def init_db() -> None:
     con = access.get_db_connection()
@@ -237,46 +231,6 @@ def build_entry_log() -> str:
     finally:
         con.close()
 
-def auth_user(pat : str) -> tuple[github.Github, AuthenticatedUser]:
-    g = github.Github(
-        auth=github.Auth.Token(pat)
-    )
-
-    user : NamedUser | AuthenticatedUser = g.get_user() # Lazy auth
-
-    # Forces a request to fetch the login
-    _ = user.login # May raise a GithubException for error status codes
-
-    assert isinstance(user, AuthenticatedUser)
-
-    return g, user
-
-def verify_repository(user : AuthenticatedUser, repo_name : str) -> Repository:
-    return user.get_repo(repo_name)
-
-def verify_permissions(repo: Repository):
-    p = repo.permissions
-
-    if not p.push and not p.pull:
-        raise MissingPermissionsError("push & pull")
-
-    if not p.push:
-        raise MissingPermissionsError("push")
-
-    if not p.pull:
-        raise MissingPermissionsError("pull")
-
-def finalise_backup_setup(pat : str, repo_name: str, user : str):
-    con = access.get_db_connection()
-
-    try:
-        set_pat(pat)
-        access.set_state(con, 'BACKUP_REPO_NAME', repo_name)
-        access.set_state(con, 'USER', user)
-        con.commit()
-    finally:
-        con.close()
-
 def get_repo_name() -> None | str:
     con = access.get_db_connection()
 
@@ -288,7 +242,7 @@ def get_repo_name() -> None | str:
 def get_state(key : str) -> None | str:
     con = access.get_db_connection()
     try:
-        return access.get_state(con, 'BACKUP_REPO_NAME')
+        return access.get_state(con, key)
     finally:
         con.close()
 
@@ -300,63 +254,6 @@ def get_user() -> None | str:
         return access.get_state(con, 'USER')
     finally:
         con.close()
-
-# TODO:Move below methods to backup.py & improve code quality
-def get_repo(auth_url,
-                        report_func: Callable[[str], None] = lambda _: None) -> git.Repo:
-    if not access.check_repo(BACKUP_REPO_DIR):
-        repo = git.Repo.clone_from(auth_url, BACKUP_REPO_DIR)
-        report_func(f"Cloned backup repo to '{BACKUP_REPO_DIR.relative_to(Path.home())}'")
-    else: 
-        repo = git.Repo(BACKUP_REPO_DIR)
-        repo.remotes.origin.set_url(auth_url)
-        report_func(f"Backup repo found '{BACKUP_REPO_DIR}'")
-    
-    return repo
-
-def populate_empty_repo(repo : git.Repo, report_func: Callable[[str], None] = lambda _ : None) -> None:
-    readme_file = BACKUP_REPO_DIR / "README.md"
-    with open(readme_file, 'w', encoding='utf-8') as f:
-        f.write("# lc-track remote backup\n Event log backup for `lc-track`")
-    repo.index.add(['README.md'])
-    repo.index.commit("Initial commit")
-    repo.description("description")
-    repo.remotes.origin.push('main:main')
-    report_func()
-
-def event_log_sync(repo : git.Repo, report_func: Callable[[str], None] = lambda _ : None) -> None:
-    repo.remotes.origin.pull()
-    report_func("Latest remote event log succesfully pulled")
-
-    event_log : list[BaseEvent] = backup.merge_event_logs(BACKUP_EVENT_LOG, LOCAL_EVENT_LOG)
-    report_func("Local and remote event logs merged")
-
-    # Atomic writes to both destinations
-    for target_path in [BACKUP_EVENT_LOG, LOCAL_EVENT_LOG]:
-        backup.write_event_log(TMP_EVENT_LOG, event_log)
-        TMP_EVENT_LOG.replace(target_path)
-
-    repo.index.add([BACKUP_EVENT_LOG.name])
-    if repo.is_dirty():
-        repo.remotes.origin.push()
-        report_func("Merged event log pushed to remote")
-    else:
-        report_func("Remote already up to date")
-    
-    backup.update_state_from_local_event_log()
-    report_func("Program state re-initialised from event log")
-
-class MissingPermissionsError(Exception):
-    pass
-
-class FailedAuthError(Exception):
-    pass
-
-def set_pat(pat : str) -> None:
-    keyring.set_password("lc-track", "gh_pat", pat)
-
-def get_pat() -> str | None:
-    return keyring.get_password("lc-track", "gh_pat")
 
 def recalc_problem_state(con : sqlite3.Connection, problem_id : int) -> None:
     entries : list[Entry] = access.get_entries_by_problem_id(con, problem_id)
@@ -388,6 +285,46 @@ def calculate_new_state(n : int, ef : float, i : int, confidence : int, now_ts :
 
     return n, ef, i, next_review_at
 
+def problem_set_sync(report_func : Callable[[str], None] = lambda _ : None) -> None:
+    raw_problem_set = lc_client.fetch_all_problems()
+    report_func("Fetched problem set from leetcode.com")
+
+    problems, topics, problem_topics = parse_raw_problem_set(raw_problem_set)
+    report_func("Parsed problem set")
+
+    con = access.get_db_connection()
+    try:
+        access.populate_db_with_problem_set(con, problems, topics, problem_topics)
+        access.set_state(con, 'initial_sync', 'complete')
+        con.commit()
+        report_func("Problem set saved")
+    finally:
+        con.close()
+
+def parse_raw_problem_set(problems_raw : list[dict[str, Any]]
+    ) -> tuple[
+        list[tuple[int, str, str, int]],
+        list[tuple[str, str]],
+        list[tuple[int, str]]]:
+    try:
+        problems = [
+            (
+                int(x['questionFrontendId']),
+                str(x['titleSlug']),
+                str(x['title']),
+                DIFF_TO_INT[x['difficulty']]
+            )
+            for x in problems_raw
+        ]
+
+        topics = list({(str(t['slug']), str(t['name'])) for p in problems_raw for t in p['topicTags']})
+
+        problem_topics = [(int(p['questionFrontendId']), str(t['slug'])) for p in problems_raw for t in p['topicTags']]
+
+        return problems, topics, problem_topics
+    except Exception as exc:
+        raise Exception("Failed to parse problem set from leetcode.com") from exc
+
 class ProblemNotFoundError(Exception):
     pass
 
@@ -399,7 +336,6 @@ class ProblemAlreadyInactiveError(Exception):
 
 class EntryNotNoundError(Exception):
     pass
-
 
 def SM2(grade : int,
         repetition_num : int,
